+++ src/components/pages/list/CheckBox.tsx
@@ -1,26 +1,23 @@
 'use client';
-
-import React, { useState } from 'react';
+import React from 'react';
 import Image from 'next/image';
 import UncheckedImage from '@/assets/icons/uncheckedbox.svg';
 import CheckedImage from '@/assets/icons/checkedbox.svg';
 
-const Checkbox: React.FC = () => {
-  const [checked, setChecked] = useState(false);
+interface CheckboxProps {
+  checked: boolean;
+  onChange: (checked: boolean) => void;
+}
 
+const Checkbox = ({ checked, onChange }: CheckboxProps) => {
   const handleChange = () => {
-    setChecked(!checked);
+    onChange(!checked); // 체크된 상태를 인자로 전달
   };
-
   return (
-    <div className="flex w-6 items-center justify-center">
-      <input
-        type="checkbox"
-        id="checkbox"
-        className="appearance-none"
-        checked={checked}
-        onChange={handleChange}
-      />
+    <div
+      className="flex w-6 items-center justify-center"
+      onClick={handleChange}
+    >
       <label htmlFor="checkbox" className="cursor-pointer">
         <Image
           src={checked ? CheckedImage : UncheckedImage}
